src/requirements.txt
Pillow==2.1.0
numpy==1.10.4
